feat (updateMDS): add API mds/update to trigger an MDS update
