--- webAuthn/validateMDS/downloadBlob.test.ts
@@ -1,9 +1,0 @@
-import {describe, expect, test} from '@jest/globals'
-import {updateMDS} from './downloadBlob'
-
-describe('Validate MDS', () => {
-  test('Update MDS to JWT', async () => {
-    await expect(updateMDS()).resolves.toBe(undefined)
-  })
-})
-
