--- webAuthn/validateMDS/index.ts
@@ -1,7 +1,0 @@
-// index file for MDS
-
-const validateMDS = () => {
-
-}
-
-export default validateMDS
